requirements.txt
omegaup
mosspy
beautifulsoup4
pybars3